requirements.txt
Flask==2.3.3
rembg==2.0.59
Pillow==10.0.1
gunicorn==21.2.0
